requirements.txt
# Flask and Web Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis and Caching
redis==5.0.1
Flask-Session==0.5.0

# Authentication and Security
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# Stripe for Billing
stripe==7.8.0

# API and AI
anthropic==0.8.0

# Validation
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2
